refactor(cli): extract reusable node launchers
