+++ bin/consensus/Cargo.toml
@@ -27,7 +27,6 @@ alloy-primitives.workspace = true
 base-common-network.workspace = true
 
 # consensus
-base-common-chains.workspace = true
 base-common-genesis.workspace = true
 base-consensus-peers.workspace = true
 base-consensus-disc = { workspace = true, features = ["metrics"] }
@@ -35,7 +34,6 @@ base-consensus-node = { workspace = true, features = ["metrics"] }
 base-consensus-providers = { workspace = true, features = ["metrics"] }
 
 # cli
-strum.workspace = true
 clap = { workspace = true, features = ["env"] }
 
 # networking
