+++ bin/node/Cargo.toml
@@ -17,25 +17,14 @@ workspace = true
 
 [dependencies]
 # workspace
-base-metering.workspace = true
 base-cli-utils.workspace = true
-base-txpool-rpc.workspace = true
-base-flashblocks.workspace = true
-base-node-runner.workspace = true
-base-tx-forwarding.workspace = true
-base-txpool-tracing.workspace = true
-base-flashblocks-node.workspace = true
-base-proofs-extension.workspace = true
-base-bundle-extension.workspace = true
 
 # reth
 base-reth-cli.workspace = true
 reth-cli-util.workspace = true
-base-node-core.workspace = true
 base-execution-cli = { workspace = true, features = ["otlp"] }
 
 # cli
-url.workspace = true
 clap.workspace = true
 
 [features]
