+++ crates/consensus/cli/src/lib.rs
@@ -16,6 +16,9 @@ pub use l1::L1ClientArgs;
 mod l2;
 pub use l2::L2ClientArgs;
 
+mod node;
+pub use node::{ConsensusNodeArgs, ConsensusNodeOverrides};
+
 mod rpc;
 pub use rpc::RpcArgs;
