+++ crates/execution/cli/Cargo.toml
@@ -11,41 +11,50 @@ repository.workspace = true
 workspace = true
 
 [dependencies]
+# reth
 reth-cli.workspace = true
 reth-discv4.workspace = true
 reth-discv5.workspace = true
 reth-net-nat.workspace = true
 reth-provider.workspace = true
-reth-node-core.workspace = true
-base-node-core.workspace = true
 reth-network.workspace = true
 reth-cli-util.workspace = true
+reth-node-core.workspace = true
+reth-chainspec.workspace = true
+reth-tracing.workspace = true
+reth-cli-runner.workspace = true
+reth-node-builder.workspace = true
 reth-network-p2p.workspace = true
 reth-cli-commands.workspace = true
+reth-node-metrics.workspace = true
 reth-network-peers.workspace = true
 reth-rpc-server-types.workspace = true
 reth-db = { workspace = true, features = ["mdbx"] }
 
-# so jemalloc metrics can be included
-reth-node-metrics.workspace = true
-
-## optimism
+# op-reth
+base-metering.workspace = true
+base-cli-utils.workspace = true
+base-node-core.workspace = true
+base-txpool-rpc.workspace = true
+base-flashblocks.workspace = true
+base-node-runner.workspace = true
 base-common-chains.workspace = true
 base-execution-evm.workspace = true
+base-tx-forwarding.workspace = true
+base-txpool-tracing.workspace = true
+base-flashblocks-node.workspace = true
 base-execution-trie.workspace = true
+base-bundle-extension.workspace = true
+base-proofs-extension.workspace = true
 base-execution-chainspec.workspace = true
 base-execution-consensus.workspace = true
-
-reth-tracing.workspace = true
-reth-chainspec.workspace = true
-reth-cli-runner.workspace = true
-reth-node-builder.workspace = true
 
 # misc
 futures.workspace = true
 clap = { workspace = true, features = ["derive", "env"] }
 
 eyre.workspace = true
+url.workspace = true
 tokio.workspace = true
 backon.workspace = true
 tracing.workspace = true
