+++ crates/execution/cli/src/standard_node.rs
@@ -1,17 +1,25 @@
-//! Contains the CLI arguments
+//! Standard Base execution-node arguments and runner wiring.
 
+use base_bundle_extension::BundleExtension;
 use base_flashblocks::FlashblocksConfig;
+use base_flashblocks_node::FlashblocksExtension;
+use base_metering::{MeteredOpcodes, MeteringConfig, MeteringExtension, MeteringResourceLimits};
 use base_node_core::args::RollupArgs;
+use base_node_runner::{BaseNodeBuilder, BaseNodeRunner, LaunchedBaseNode};
+use base_proofs_extension::ProofsHistoryExtension;
 use base_tx_forwarding::{
     DEFAULT_MAX_BATCH_SIZE, DEFAULT_MAX_RPS, DEFAULT_RESEND_AFTER_MS, TxForwardingConfig,
+    TxForwardingExtension,
 };
+use base_txpool_rpc::{TxPoolRpcConfig, TxPoolRpcExtension};
+use base_txpool_tracing::{TxPoolExtension, TxpoolConfig};
 use url::Url;
 
-/// CLI Arguments
+/// CLI arguments for a standard Base execution node.
 #[derive(Debug, Clone, PartialEq, Eq, clap::Args)]
 #[command(next_help_heading = "Rollup")]
-pub struct Args {
-    /// Rollup arguments
+pub struct StandardNodeArgs {
+    /// Rollup arguments.
     #[command(flatten)]
     pub rollup_args: RollupArgs,
 
@@ -131,8 +139,8 @@ pub struct Args {
     pub tx_forwarding_max_rps: u32,
 }
 
-impl From<&Args> for Option<FlashblocksConfig> {
-    fn from(args: &Args) -> Self {
+impl From<&StandardNodeArgs> for Option<FlashblocksConfig> {
+    fn from(args: &StandardNodeArgs) -> Self {
         args.flashblocks_url.clone().map(|url| {
             let mut config = FlashblocksConfig::new(url, args.max_pending_blocks_depth);
             config.cached_execution = args.flashblocks_cached_execution;
@@ -141,8 +149,8 @@ impl From<&Args> for Option<FlashblocksConfig> {
     }
 }
 
-impl From<&Args> for TxForwardingConfig {
-    fn from(args: &Args) -> Self {
+impl From<&StandardNodeArgs> for TxForwardingConfig {
+    fn from(args: &StandardNodeArgs) -> Self {
         if !args.enable_tx_forwarding || args.builder_rpc_urls.is_empty() {
             return Self::default();
         }
@@ -151,5 +159,82 @@ impl From<&Args> for TxForwardingConfig {
             .with_resend_after_ms(args.tx_forwarding_resend_after_ms)
             .with_max_batch_size(args.tx_forwarding_batch_size)
             .with_max_rps(args.tx_forwarding_max_rps)
+    }
+}
+
+/// Standard Base execution-node runner wiring.
+#[derive(Debug, Clone, Copy)]
+pub struct StandardBaseRethNode;
+
+impl StandardBaseRethNode {
+    /// Builds a runner with the standard Base execution-node extensions installed.
+    pub fn runner(args: StandardNodeArgs) -> eyre::Result<BaseNodeRunner> {
+        let mut runner = BaseNodeRunner::new(args.rollup_args.clone());
+
+        // Create flashblocks config first so we can share its state with metering.
+        let flashblocks_config: Option<FlashblocksConfig> = (&args).into();
+
+        // Feature extensions (FlashblocksExtension must be last - uses replace_configured).
+        runner.install_ext::<TxPoolRpcExtension>(TxPoolRpcConfig {
+            sequencer_rpc: args.rollup_args.sequencer.clone(),
+        });
+        runner.install_ext::<TxPoolExtension>(TxpoolConfig {
+            tracing_enabled: args.enable_transaction_tracing,
+            tracing_logs_enabled: args.enable_transaction_tracing_logs,
+            flashblocks_config: flashblocks_config.clone(),
+        });
+
+        let resource_limits = MeteringResourceLimits {
+            gas_limit: args.metering_gas_limit,
+            execution_time_us: args.metering_execution_time_us,
+            state_root_time_us: args.metering_state_root_time_us,
+            da_bytes: args.metering_da_bytes,
+        };
+        let metering_config = if args.enable_metering {
+            let metered_opcodes = if args.metering_metered_opcodes.is_empty() {
+                MeteredOpcodes::default()
+            } else {
+                MeteredOpcodes::parse(&args.metering_metered_opcodes)?
+            }
+            .with_all_precompiles();
+
+            let mut config = flashblocks_config
+                .clone()
+                .map_or_else(MeteringConfig::enabled, MeteringConfig::with_flashblocks)
+                .with_resource_limits(resource_limits)
+                .with_metered_opcodes(metered_opcodes);
+            if let Some(target_flashblocks_per_block) = args.metering_target_flashblocks_per_block {
+                config = config.with_target_flashblocks_per_block(target_flashblocks_per_block);
+            }
+            config
+        } else {
+            MeteringConfig::disabled()
+        };
+        runner.install_ext::<MeteringExtension>(metering_config);
+        runner.install_ext::<BundleExtension>(());
+        runner.install_ext::<TxForwardingExtension>((&args).into());
+        runner.install_ext::<FlashblocksExtension>(flashblocks_config);
+        runner.install_ext::<ProofsHistoryExtension>(args.rollup_args);
+
+        Ok(runner)
+    }
+
+    /// Launches the node and waits for it to exit.
+    pub async fn run(builder: BaseNodeBuilder, args: StandardNodeArgs) -> eyre::Result<()> {
+        let mut runner = Self::runner(args)?;
+        runner.add_started_callback(|| {
+            base_cli_utils::register_version_metrics!();
+            Ok(())
+        });
+
+        runner.run(builder).await
+    }
+
+    /// Launches the node and returns immediately with a handle.
+    pub async fn launch(
+        builder: BaseNodeBuilder,
+        args: StandardNodeArgs,
+    ) -> eyre::Result<LaunchedBaseNode> {
+        Self::runner(args)?.launch(builder).await
     }
 }
