+++ crates/execution/runner/src/runner.rs
@@ -17,6 +17,13 @@ use crate::{
 
 type StartedCallback = Box<dyn FnOnce() -> Result<()> + Send + 'static>;
 
+/// Handle to a launched Base execution node.
+#[derive(Debug)]
+pub struct LaunchedBaseNode {
+    /// The underlying reth node handle.
+    pub handle: NodeHandleFor<BaseNode>,
+}
+
 /// Wraps the Base node configuration and orchestrates builder wiring.
 pub struct BaseNodeRunner<SB: PayloadServiceBuilder = DefaultPayloadServiceBuilder> {
     /// Rollup-specific arguments forwarded to the Base node implementation.
@@ -89,8 +96,17 @@ impl<SB: PayloadServiceBuilder> BaseNodeRunner<SB> {
     /// Applies all Base-specific wiring to the supplied builder, launches the node, and waits for
     /// shutdown.
     pub async fn run(self, builder: BaseNodeBuilder) -> Result<()> {
+        let LaunchedBaseNode { handle: NodeHandle { node: _node, node_exit_future } } =
+            self.launch(builder).await?;
+        node_exit_future.await?;
+        Ok(())
+    }
+
+    /// Applies all Base-specific wiring to the supplied builder and returns a launched node
+    /// handle without waiting for shutdown.
+    pub async fn launch(self, builder: BaseNodeBuilder) -> Result<LaunchedBaseNode> {
         let Self { rollup_args, extensions, service_builder, da_config, started_callbacks } = self;
-        let NodeHandle { node: _node, node_exit_future } = Self::launch_node(
+        let handle = Self::launch_node(
             rollup_args,
             extensions,
             service_builder,
@@ -99,8 +115,7 @@ impl<SB: PayloadServiceBuilder> BaseNodeRunner<SB> {
             builder,
         )
         .await?;
-        node_exit_future.await?;
-        Ok(())
+        Ok(LaunchedBaseNode { handle })
     }
 
     async fn launch_node(
